+++ src/test/kotlin/com/taskforce/superinvention/common/aws/s3/AwsS3MoTest.kt
@@ -2,8 +2,8 @@ package com.taskforce.superinvention.common.aws.s3
 
 import com.taskforce.superinvention.common.util.aws.s3.AwsS3Mo
 import com.taskforce.superinvention.config.test.IntegrationTest
-import org.junit.Ignore
 import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.mock.web.MockMultipartFile
@@ -22,8 +22,8 @@ class AwsS3MoTest: IntegrationTest() {
         multipartFile = MockMultipartFile("test-image-file", "test-image-file.png", "multipart/form-data", inputFile)
     }
 
-    @Ignore
-    @Test
+
+    @Disabled @Test
     fun `AWS S3 파일 업로드`() {
         val s3Path = awsS3Mo.uploadFile(multipartFile, "temp")
         print(s3Path)
