+++ src/test/kotlin/com/taskforce/superinvention/document/club/ClubDocumentation.kt
@@ -46,96 +46,6 @@ class ClubDocumentation: ApiDocumentationTest() {
 
     @Test
     @WithMockUser(authorities = [Role.MEMBER])
-    fun `모임 상세 조회`() {
-
-        // given
-        val user = User ("sight")
-        val club = Club(
-                name = "가상 모임",
-                description = "가상 모임에 대한 설명",
-                maximumNumber = 100L,
-                mainImageUrl = ""
-        )
-
-        val clubUser = ClubUser(club, user, isLiked = false)
-
-        user.seq = 1
-        club.seq = 2
-        clubUser.seq = 3
-
-        val region1 = Region(
-                superRegion = null,
-                name = "성남시",
-                superRegionRoot = "경기도/성남시",
-                level = 2,
-                subRegions = listOf()
-        )
-        region1.seq = 401
-
-        val clubRegionList = listOf(
-                SimpleRegionDto(seq = 101, name = "강남구", superRegionRoot = "서울특별시/강남구", level = 2)
-        )
-
-        val interestList = listOf(
-                InterestWithPriorityDto(InterestDto(11, "등산", SimpleInterestGroupDto(20, "운동/건강")), 2)
-        )
-
-        val resultDto = ClubInfoDetailsDto (
-                clubInfo = ClubInfoDto(
-                        club = club,
-                        userCount = 20L,
-                        clubInterest = interestList,
-                        clubRegion   = clubRegionList
-                ),
-                userInfo = ClubUserStatusDto(
-                        role = listOf(Role.RoleName.CLUB_MEMBER, Role.RoleName.MANAGER),
-                        isLiked = false
-                )
-        )
-
-        // when
-        `when`(clubService.getClubInfoDetail(MockitoHelper.anyObject(), eq(club.seq!!))).thenReturn(resultDto)
-
-        val result = mockMvc.perform(
-                get("/clubs/{clubSeq}", club.seq)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .accept(MediaType.APPLICATION_JSON)
-                        .header("Authorization", "Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdXRoIjoiW1VTRVJdIi")
-                        .characterEncoding("UTF-8")
-        ).andDo(print())
-
-        result.andExpect(status().isOk)
-                .andDo(
-                        document("select-club-info-detail",
-                                getDocumentRequest(),
-                                getDocumentResponse(),
-                                pathParameters(parameterWithName("clubSeq").description("모임 시퀀스.")),
-                                responseFields(
-                                        *commonResponseField(),
-                                        fieldWithPath("data.clubInfo.seq").type(JsonFieldType.NUMBER).description("모임 시퀀스"),
-                                        fieldWithPath("data.clubInfo.name").type(JsonFieldType.STRING).description("모임 명"),
-                                        fieldWithPath("data.clubInfo.description").type(JsonFieldType.STRING).description("모임 설명"),
-                                        fieldWithPath("data.clubInfo.maximumNumber").type(JsonFieldType.NUMBER).description("모임 최대 가입 인원 수"),
-                                        fieldWithPath("data.clubInfo.userCount").type(JsonFieldType.NUMBER).description("현재 모임원 인원수"),
-                                        fieldWithPath("data.clubInfo.mainImageUrl").type(JsonFieldType.STRING).description("모임 대표 사진"),
-                                        fieldWithPath("data.clubInfo.clubInterest[].interest.seq").type(JsonFieldType.NUMBER).description("모임 관심사 seq"),
-                                        fieldWithPath("data.clubInfo.clubInterest[].interest.name").type(JsonFieldType.STRING).description("모임 관심사명"),
-                                        fieldWithPath("data.clubInfo.clubInterest[].interest.interestGroup.seq").type(JsonFieldType.NUMBER).description("모임 관심사 그룹 Seq"),
-                                        fieldWithPath("data.clubInfo.clubInterest[].interest.interestGroup.name").type(JsonFieldType.STRING).description("모임 관심사 그룹명"),
-                                        fieldWithPath("data.clubInfo.clubInterest[].priority").type(JsonFieldType.NUMBER).description("모임 관심사 우선순위"),
-                                        fieldWithPath("data.clubInfo.clubRegion[].seq").type(JsonFieldType.NUMBER).description("모임 지역 seq"),
-                                        fieldWithPath("data.clubInfo.clubRegion[].name").type(JsonFieldType.STRING).description("모임 지역명"),
-                                        fieldWithPath("data.clubInfo.clubRegion[].superRegionRoot").type(JsonFieldType.STRING).description("모임 상위 지역"),
-                                        fieldWithPath("data.clubInfo.clubRegion[].level").type(JsonFieldType.NUMBER).description("모임 지역 단계"),
-                                        fieldWithPath("data.userInfo.role[]").type(JsonFieldType.ARRAY).description("유저 권한"),
-                                        fieldWithPath("data.userInfo.isLiked").type(JsonFieldType.BOOLEAN).description("모임원 모임 좋아요 여부")
-                                )
-                        )
-                )
-    }
-
-    @Test
-    @WithMockUser(authorities = [Role.MEMBER])
     fun `모임 생성`() {
         val clubAddRequestDto = ClubAddRequestDto(
                 name = "땔감 스터디",
